Projects/requirements.txt
numpy
Flask
gunicorn
sklearn
pandas
PyYAML
-e .

